+++ hamcrest-matchers/java7-hamcrest-matchers/src/test/java/it/ozimov/cirneco/hamcrest/java7/AssertionMakerTest.java
@@ -1,0 +1,41 @@
+package it.ozimov.cirneco.hamcrest.java7;
+
+import static org.junit.Assert.*;
+
+import org.junit.Test;
+
+public class AssertionMakerTest {
+
+    @Test
+    public void testGiven() throws Exception { }
+
+    @Test
+    public void testWithReason() throws Exception { }
+
+    @Test
+    public void testForReason() throws Exception { }
+
+    @Test
+    public void testAssumeThat() throws Exception { }
+
+    @Test
+    public void testAssumeIs() throws Exception { }
+
+    @Test
+    public void testAssertThat() throws Exception { }
+
+    @Test
+    public void testAssertIs() throws Exception { }
+
+    @Test
+    public void testAssumeThat1() throws Exception { }
+
+    @Test
+    public void testAssumeIs1() throws Exception { }
+
+    @Test
+    public void testAssertThat1() throws Exception { }
+
+    @Test
+    public void testAssertIs1() throws Exception { }
+}